services/admin/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
jinja2==3.1.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
redis==5.0.1
bcrypt==4.1.2
python-dotenv==1.0.0
httpx==0.26.0
